requirements.txt
streamlit
pandas
altair
gspread
google-auth
google-auth-oauthlib
python-dotenv
requests
google-generativeai
certifi
